requirements.txt
alabaster==0.7.9
appdirs==1.4.2
Babel==2.3.4
cffi==1.9.1
cryptography==1.7.2
Django==1.10.6
-e git+https://github.com/qdqmedia/django-actions.git@76750bf815aa7c3de3e65506d4b78576f92b1287#egg=django_actions
django-bootstrap-themes==3.3.6
django-cas-ng==3.5.6
django-collectd-rest==0.1.9.1
django-crispy-forms==1.6.1
django-geojson==2.9.1
django-guardian==1.4.6
django-leaflet==0.19.0
django-model-utils==2.6.1
django-netjsonconfig>=0.5.3.post1
django-reversion==2.0.8
django-search-listview==0.1.1
django-sortedm2m==1.3.3
django-x509==0.2
djangorestframework==3.5.3
docutils==0.12
idna==2.4
imagesize==0.7.1
Jinja2==2.9.5
jsonfield==2.0.0
jsonschema==2.6.0
MarkupSafe==0.23
netjsonconfig==0.5.4
packaging==16.8
pyasn1==0.2.3
pycparser==2.17
Pygments==2.1.3
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-cas==1.2.0
pytz==2016.7
requests==2.12.4
rules==1.1.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.8
uWSGI==2.0.14
